+++ docs/md/manual/releasenotes.md
@@ -11,6 +11,7 @@
 - Fix `-keepparameternames` to keep Kotlin interface parameter names.
 - Fix potential `NullPointerException` while processing enum classes with invalid Kotlin metadata.
 - Fix potential `Instruction has invalid constant index size` error during GSON optimization.
+- Fix member specialization & generalization optimizations.
 
 ## Version 7.3.0